requirements.txt
numpy
opencv-python
pillow
openai
customtkinter>=5.2.0
pywin32
keyboard
faster-whisper
pyaudio
